+++ changelog.md
@@ -1,5 +1,15 @@
 
-[v0.28.1](https://github.com/nextcloud/nextcloudpi/commit/164b656) (2017-09-15) nc-static-IP: add new IP to trusted domain
+[v0.29.3](https://github.com/nextcloud/nextcloudpi/commit/93c19f2) (2017-09-21) docker: updated to stretch and added NextCloudPi layer
+
+[v0.29.2 ](https://github.com/nextcloud/nextcloudpi/commit/94a6513) (2017-09-20) changed hostname
+
+[v0.29.1 ](https://github.com/nextcloud/nextcloudpi/commit/03b0263) (2017-09-20) lamp: add fileinfo mcrypt packages
+
+[v0.29.0 ](https://github.com/nextcloud/nextcloudpi/commit/74a97c3) (2017-09-20) updated to NC12.0.3
+
+[v0.28.2](https://github.com/nextcloud/nextcloudpi/commit/ac3fcab) (2017-09-18) ncp-web: point changelog to master
+
+[v0.28.1 ](https://github.com/nextcloud/nextcloudpi/commit/3b71b3d) (2017-09-15) nc-static-IP: add new IP to trusted domain
 
 [v0.28.0 ](https://github.com/nextcloud/nextcloudpi/commit/6cf07b2) (2017-09-14) added nc-static-IP